+++ cmd/tether/attach.go
@@ -22,12 +22,14 @@ import (
 	"sync"
 	"sync/atomic"
 	"syscall"
+	"time"
 
 	log "github.com/Sirupsen/logrus"
 	"golang.org/x/crypto/ssh"
 
 	"github.com/vmware/vic/cmd/tether/msgs"
 	"github.com/vmware/vic/lib/tether"
+	"github.com/vmware/vic/pkg/serial"
 	"github.com/vmware/vic/pkg/trace"
 )
 
@@ -218,6 +220,42 @@ func (t *attachServerSSH) stop() error {
 	log.Debugf("Released the connection lock")
 
 	return nil
+}
+
+func backchannel(ctx context.Context, conn *net.Conn) error {
+	defer trace.End(trace.Begin("establish tether backchannel"))
+
+	// HACK: currently RawConn dosn't implement timeout so throttle the spinning
+	// it does implement the Timeout methods so the intermediary code can be written
+	// to support it, but they are stub implementation in rawconn impl.
+
+	// This needs to tick *faster* than the ticker in connection.go on the
+	// portlayer side.  The PL sends the first syn and if this isn't waiting,
+	// alignment will take a few rounds (or it may never happen).
+	ticker := time.NewTicker(10 * time.Millisecond)
+	defer ticker.Stop()
+
+	ch := make(chan struct{}, 1)
+
+	go func() {
+		select {
+		case <-ctx.Done():
+			(*conn).Close()
+			close(ch)
+		}
+	}()
+
+	for {
+		select {
+		case <-ticker.C:
+			err := serial.HandshakeServer(ctx, *conn)
+			if err == nil {
+				return nil
+			}
+		case <-ch:
+			return ctx.Err()
+		}
+	}
 }
 
 // run should not be called directly, but via start
